#44 improvement disk checker.
