+++ src/main/java/by/ladyka/club/service/disk/SpaceDiskService.java
@@ -1,4 +1,4 @@
-package by.ladyka.club.service;
+package by.ladyka.club.service.disk;
 
 public interface SpaceDiskService {
     void check();
